Feat: useGame 커스텀훅 분리
